upgrade to docker-client:4.0.8
This latest version of docker-client includes a change in the
`DefaultDockerClient.fromEnv()` so that on OS X it uses
`unix:///var/run/docker.sock` by default if DOCKER_HOST is not set (to
work with Docker for Mac out of the box).

In 4.x of the docker-client the package names of exception classes
changed as well.